chore: solved day03
Part 2 took considerably longer because I did not realise the example input had
changed. In fact, part 1 used underscores (`_`) instead of single quotes (`'`).

But a regex solved these problems nicely. For part 2, you also had to be careful
not to change the `mul_enabled` while there was first a `mul()` operation (which
is why the extra checks for `cap.start() == 0` are needed).

```txt
    Time  Rank  Score       Time  Rank  Score
00:05:01  1518      0   00:22:34  4233      0
```