chore: solved day04
Got stuck on part 2 again, as I tried to solve it with the directionality checks
from part 1. However, each time I ended up with too few `X-MAS` matches and
making the checks more complex and adding more and more edge cases was making
it difficult to see what was going wrong in the 50+ LOC at that point.

So went back to the drawing board after a good cup of hot chocolate and came up
with something that was a lot simpler.

```txt
    Time  Rank  Score       Time  Rank  Score
00:15:36  3007      0   00:52:08  6683      0
```