chore: solved day05
Lots of fun! Less input parsing puzzles is more better!

Somehow, my solution for part 2 always ends up too complicated until I take a
step back to look at the problem again. I started with implementing a completely
custom sorting stack to manage the relations in the page ordering rules.

Initially, I got stuck at the code being non-deterministic. This was,
unfortunately, the result of trying to copy the `update` into a `HashSet`. Doing
this cascaded down to the selection of the next element to process, where it
thus could happen that a random element was selected which could never produce
a correct _updated_ ordering.

No hot chocolate this morning, but drawing out on paper what you are trying to
do is still the best way to solve these puzzles. At that point I dropped the
custom implementation that kept track of both directions in the page ordering
rules and went for a simpler direct sorting approach.

```txt
    Time  Rank  Score       Time  Rank  Score
00:11:32  1706      0   00:37:41  4131      0
```