chore: solved day08
Did you know that knowing your bounds is very important, and that off-by-one
errors are frustrating to deal with? And that if you have managed to produced
such a mistake it can cost you up to **20** minutes to figure out your mistake?

:'(

This also required an unfortunate rewrite of part 1 for part 2 as it could
initially not be integrated into what I had for part 1. However, the result is
code and logic that is pretty clean (what started as something with 3 case
distinctions has no case distinctions now).

```txt
    Time  Rank  Score       Time  Rank  Score
00:40:21  5417      0   00:59:43  5691      0
```