chore: solved day09
Fun! Made plenty of mistakes, by creating an initial implementation that would
forget the last file if there were free spaces after it. Most certainly made a
shifting function that went right-to-left instead of left-to-right (because
reading is difficult).

Already includes a few minor touch-ups after getting part 2 to make the handling
of file sizes (especially for the blocks in part 2) more manageable. This means
that the actual disk compacting function does not need to know the difference
between part 1 and part 2, as this is already handled at the input parsing!

```txt
    Time  Rank  Score       Time  Rank  Score
00:40:59  5244      0   00:47:08  1687      0
```

(so with the extra touch-ups and less spaghetti for part 2, this took a bit more
than an hour instead of the 47 minutes)