chore: solved day21
I cannot function on less than 3 hours of sleep...

This was incredibly painful; started with BFS but could not get it to work in
any way. Switched to a more logical approach that just follows the Manhattan
distance. However, it is important to choose the specific path to be able
to optimise subsequent iterations on that path as a new code for the next robot.

I learned a lot about generics in Rust! Because with the two keypads having
slightly different sizes I did not want to write each function for both of them.

I am going back to sleep and I hope I wake up just before 06:00 tomorrow to be
ready for the next puzzle.

```
    Time  Rank  Score       Time  Rank  Score
01:30:46   975      0   02:12:03   756      0
```