chore: solved day22
No graph this morning makes this a good morning (it was already a good morning
because I have not felt this well-rested in a few days).

Tried a few approaches for part 2 that ended up being too slow, but building up
the sequence of ones digits already while generating the secret numbers helps a
lot (because I did this first after was unnecessarily slow). Apart from that
there is not really anything special about my solution (just checks each pattern
and update the total banana count for each unique pattern).

```txt
    Time  Rank  Score       Time  Rank  Score
00:05:24   448      0   00:44:30  1541      0
```